Changed app design and structure
Renamed ViewController.swift to PeersViewController.
Created the main view controller, MessagesViewController, which will contain all of the messages sent and recieved.
In MessagesViewController.swift, added an indicator indicating how many peers are connected to the client indirectly, and how many peers are directly connected. Next to it, added another indicator indicating if the client is connected to the server.
In SocketServiceManager.swift, added the getConnectedPeers function for the indicator in the messages view controller.
Created NavigationController.swift for when changes to the navigation controller will be necessary.